+++ src/test/java/org/eocqrs/kafka/data/KfDataTest.java
@@ -20,33 +20,47 @@
  * SOFTWARE.
  */
 
-package org.eocqrs.kafka;
+package org.eocqrs.kafka.data;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import com.jcabi.xml.XMLDocument;
-import java.io.File;
-import java.io.FileNotFoundException;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.eocqrs.kafka.producer.KfProducerSettings;
+import org.eocqrs.kafka.Data;
+import org.eocqrs.kafka.data.KfData;
 import org.junit.jupiter.api.Test;
 
 /**
- * Test case for {@link KfProducerSettings}
+ * Test case for {@link KfData}
  *
  * @since 0.0.0
  */
-class KfProducerSettingsTest {
+class KfDataTest {
 
   @Test
-  void testProducerConstruction() throws FileNotFoundException {
-    final ProducerSettings<String, String> settings =
-      new KfProducerSettings<>(
-        new XMLDocument(
-          new File("src/test/resources/settings.xml")
-        )
-      );
-    final KafkaProducer<String, String> out = settings.producer();
-    assertThat(out).isNotNull();
+  void stringDataized() {
+    final String origin = "test";
+    final Data<String> data = new KfData<>(
+      origin, "processes", 3
+    );
+    assertDoesNotThrow(data::dataized);
+    assertThat(data.dataized()).isNotNull();
+  }
+
+  @Test
+  void topic() {
+    final String topic = "tx-log";
+    final Data<String> data = new KfData<>(
+      "test", topic, 12
+    );
+    assertThat(data.topic()).isEqualTo(topic);
+  }
+
+  @Test
+  void partition() {
+    final int partition = 1;
+    final Data<String> data = new KfData<>(
+      "test", "customer-messages", partition
+    );
+    assertThat(data.partition()).isEqualTo(partition);
   }
 }
